playground/portal-offboarding-poc/requirements.txt
configargparse
psycopg2-binary
python-dotenv